Change IEEEReal.setRoundingMode to throw exception if setting the mode fails
As per
http://www.standardml.org/Basis/ieee-float.html#SIG:IEEE_REAL.setRoundingMode:VAL
IEEEReal.setRoundingMode should probably throw an exception if
the hardware doesn't support it. This is common on ARM, which is only
guaranteed to support the default mode.

This throws PosixError.inval, as suggested by
http://mlton.org/pipermail/mlton/2009-April/030535.html

The MLB file had to be reordered a bit to define the exception
to be raised. ./bin/regression shows no regressions on x86_64.

SVN r7082