Simplify upgrade-basis script for MLton.Platform.
upgrade-basis.sml is only used when compiling with stubs and
<src>/lib/stubs/mlton-stubs/mlton.sml handles converting from a
"simple" MLton.Platform.{Arch,OS}.t type to the
MLton.Platform.{Arch,OS}.t datatypes.  This minimizes the number of
places where MLton.Platform structures need to be kept in sync.

However, since upgrade-basis.sml is only used when compiling with
stubs, it isn't clear why it should fake the host for the purposes of
cross-compiling.  It seems more appropriate to cross-compile with a
non-mlton-stubs executable and with non-mlton-stubs sources.

SVN r7129