use-mmap true is the new default.
All regressions except socket and textio.2 pass on x86-cygwin.  The
former is probably a Windows firewall issue; the latter is due to
behind-the-scenes CR/LF <=> LF conversions.  Also, echo raises a
Windows firewall allert, but passes.  (Oddly, it continues to pass
even if you tell Windows to block the application.)



SVN r7182