Where there is a "to", there should be a "from".
Converting an intinf to/from a byte-stream is a pretty important operation 
in cryptography. The basis library doesn't provide a way to convert to bytes
aside from >> and WordX.fromLargeInt. It is not possible to convert an integer
to bytes in linear time with these operations. Therefore, this patch extends
the MLton extension to converts IntInf to/from it's underlying representation.



SVN r7223