I know I said hurd seemed like it would be hard work, but ... Pretty much all that needed to be done was add it to the OS datatype! If you find this patch too intrusive, revert it.
However, the regressions mostly all pass (no mlprof) and it is a debian 
architecture so I'd like to build it. A broken port is better than no port.

To be frank, the biggest problem with running the regressions is that the
hurd kernel keeps crashing! ... not the regressions. ;p



SVN r7287