Fix another (latent) performance bug in simplifyTypes.
Although not observed in practice, this is another case where a
structural walk over the entire type may repeatedly visit the same
tuple node with no additional effect.  As before, replace the looping
implementation with a memoizing implementation (which visits each type
exactly once).  It is possible to observe the performance bug by
performing a second simplifyTypes optimization pass immediately after
the first, with the source program that demonstrated the performance
bug fixed in r7377.

SVN r7378