Fix performance bug in refFlatten.
In computing the set of (mutually) recursive datatypes, a dependency
graph is constructed with nodes corresponding to tycons and edges
corresponding to use of (another) tycon in the constructors of a
tycon.  Similar to r7377 and r7378, this is a case where a structural
walk over an entire type may repeatedly visit the same object node
with no additional effect.  As before, replace the looping
implementation with a memoizing implementation (which visits each type
exactly once).

SVN r7379