Fix bug with type checking of datatype declarations with withtype bindings.
The type checking and elaboration of datatype declarations had been
implemented by repeatedly elaborating the withtype bindings (in the
scope of the datatype type constructors) followed by the datatype
bindings (constructors) until the equality status of the datatype
tycons stabilized.

This implementation, however, has a problem with the following:
  type u = real ;
  datatype t = A of v | B of u
  withtype u = int and v = u ;

This should elaborate as:
  type u = real;
  datatype t = A of real | B of int ;
  type u = int and v = real ;
where the tycon t does not have equality status.

It was mistakenly elaborating as:
  type u = real ;
  datatype t = A of int | B of int ;
  type u = int and v = int ;
where the tycon t does not have equality status.  This was due to the
fact that after the first round of elaboration, the tycon t changed
from having equality status to not having equality status.  This
triggered a second round of elaboration, but in an environment where
"type u = int and v = real".  Thus, in the second round of
elaboration, the withtype bindings elaborate to
"type u = int and v = int" and the constructors elaborate to
"A of int" and "B of int".  However, the monotonicity of equality
status changes meant that the tycon t remained as not having equality
status.

To implement proper type checking of datatype declarations with
withtype bindings, elaborate the bindings exactly once and make the
determination of equality status from a tycon application "lazy", and
therefore sensitive to the changes in the equality status of the
tycon.

SVN r7434