Don't build basis-ffi.sml and basis-ffi.h when building from clean sources.
It is convenient (for cross-compiling and porting) for the runtime to
build without requiring a working mlton.  Since basis-ffi.def changes
rarely, both basis-ffi.sml and basis-ffi.h are kept in the source
repository.  Also keep the gen-basis-ffi.stamp file (used to support
the multiple target Makefile rule) in the source repository.



SVN r7447