Replaced use of memcpy by memmove in mark-compact garbage collector.
Fixed bug in the mark-compact garbage collector where the C library's
memcpy was used to move objects during the compaction phase.  Objects
could be moved by an arbitrarily small amount, leading to cases where
the object's destination overlapped with the object's source.  This
could lead to heap corruption and segmentation faults with newer
versions of gcc and/or glibc, which assume that src and dst in a
memcpy do not overlap.

SVN r7461