Cast offsetof and sizeof GC constants to Word32.
The various GC constants are emitted by code like:
  fprintf (stdout, "stackTop_Offset = %"PRIu32"\n", offsetof (struct GC_state, stackTop));
The constants are known to be small, so Word32 suffices to represent
the offsets and sizes.  However, on a 64-bit platform, offsetof and
sizeof return a 64-bit unsigned integer.  Although not observed in
practice, an ABI might have sufficiently different calling conventions
for 32-bit and 64-bit values that passing a 64-bit value where a
varargs function expects a 32-bit value would not simply see the low
32-bits of the 64-bit value.

The simple solution is to cast the offsetof and sizeof expressions to
Word32 before passing them to fprintf.

SVN r7464