Fix bug in comment-handling in lexer for mlyacc's input language.
[Submitted by Michael Norrish.]

In particular, if a comment was successfully lexed, it returned a
BOGUS_VALUE token, which then messed up the higher-level parsing.
This was due to the use of the idiom

  continue() before YYBEGIN SOMESTATE

making the return value of the continue the return value for the lexer
at that point.

This bug prevented the use of comments in code sections entirely.

It was somewhat masked in the prefix section by the fact that most
supposed comments there were never lexed as such anyway.  This could
be observed by a file that included %% in a comment in the prefix
section.



SVN r7500