Fixed bug in translation from SSA2 to RSSA with weak pointers.
A weak pointer to a non-heap allocated object (e.g., an int) is always
nulled.  For such weak pointers, the translation from SSA2 to RSSA
eliminates the weak pointer operations.  However, this could leave
unbound variables.  For example, the SSA2 fragment

  L_311 ()
    x_701: word32 = Weak_get (x_697)
    x_700: bool = Weak_canGet (x_697)
    case x_700 of
      true => L_313 | false => L_312
  L_313 ()
    L_242 (x_701)

is translated to the RSSA fragment:

 L_311 () Jump =
    x_700: Word32 = 0x0: Word32
    switch {test = x_700, default = None, cases = ((0x0, L_1356), (0x1, L_1357))}
  L_313 () Jump =
    L_242 (x_701)

Although the L_313 block (and, therefore, the reference to x_701) is
dead code, the fragment is nonetheless illegal.

The fix is straighforward: rather than elide the Weak_get operation in
this scenario, replace it with a bogus value.  Thus, the above SSA2
fragment is now translated to the RSSA fragment:

 L_311 () Jump =
    x_701: Word32 = 0x1: Word32
    x_700: Word32 = 0x0: Word32
    switch {test = x_700, default = None, cases = ((0x0, L_1356), (0x1, L_1357))}
  L_313 () Jump =
    L_242 (x_701)

Thanks to Alexandre Hamez for the bug report.

SVN r7502