Unify SSA/SSA2 type checking of case expressions over words and cons.
Previously, checking for exhaustiveness and redundancy of a case
exprssion over cons used a boolean array of length equal to the number
of cons in the datatype of the test.  This had a latent "bug" in that
an ill-typed program might have a con from a different datatype in a
case expression; this con might have an index larger than the number
of cons in the datatype of the test, which would trigger an unhelpful
Subscript exception.  Verifying that the cons are appropriate for the
datatype of the test is handled by analyze, which runs after
checkScopes (which performs the exhaustiveness and redundancy checks).

SVN r7543