Fixed bug in SSA/SSA2 shrinker.
Fixed bug in SSA/SSA2 shrinker that could erroneously turn a non-tail
function call with a Bug transfer as its continuation into a tail
function call. The bug was triggered by the following SSA fragment:

fun tuple_1213 (env_9399: (lambdas_9377 * lambdas_9378 * lambdas_2366 * lambdas_234),
               x_108593: Region.Wrap.t_8 vector): {raises = Some (CodeGen.AMD64MLTree.an),
                                                   returns = Some ()} = L_39443 ()
  ...
  L_39452 (env_9400: lambdas_161 ref)
    sub_288 (env_9400, x_108600) NonTail {cont = L_39453,
                                          handler = Handle L_39454}
  L_39454 (x_108601: CodeGen.AMD64MLTree.an)
    Leave AstCore.Exp.tuple mlton/ast/ast-core.fun 625.11
    raise (x_108601)
  L_39453 (x_108602: Region.Wrap.t_8)
    region_7 (x_108602)
  region_7 (x_306671: Region.Wrap.t_8)
    L_161558 ()
  L_161558 ()
    Bug
  ...

which was transformed to:

fun tuple_1213 (env_9399: (lambdas_9377 * lambdas_9378 * lambdas_2366 * lambdas_234),
               x_108593: Region.Wrap.t_8 vector): {raises = Some (CodeGen.AMD64MLTree.an),
                                                   returns = Some ()} = L_39443 ()
  ...
  L_39452 (env_9400: lambdas_161 ref)
    Leave AstCore.Exp.tuple mlton/ast/ast-core.fun 625.11
    sub_288 (env_9400, x_108600) Tail
  ...

Note that sub_288 returns Region.Wrap.t_8, but tuple_1213 returns ();
hence the transformed program is ill-typed.

The shrinker attempts to turn a nontail call into a tail call when the
return continuation is a Bug transfer and the handler is Leaves
followed by a Raise. To enable this transformation, the shrinker only
assigns LabelMeaning.Bug to blocks where the formals match the
function return, the statements are exclusively profile statements,
and the transfer is Bug. However, the shrinker also assigned
LabelMeaning.Bug to blocks where the statements are exclusively
profile statements and the transfer is a Goto to a block assigned
LabelMeaning.Bug. Hence, in the above, LabelMeaning.Bug is assigned to
L_161558, region_7, and L_39453.

The fix is to not propagate the LabelMeaning.Bug to a block with a
Goto transfer unless the block's formals match the function return.

Thanks to Lars Bergstrom for the bug report.

SVN r7546