Fixed bug in elaboration of type variables with and without equality status.
Thanks to Rob Simmons for the bug report and examples.

Fixed bug in elaboration that erroneously rejected the following:
  datatype ('a, ''a) t = T
  type ('a, ''a) u = unit
and erroneously accepted the following:
  fun f (x: 'a) : ''a = x
  fun g (x: 'a) : ''a = if x = x then x else x

Changed ast/tyvar.fun to record full name of type variable (including
the '- or ''-prefix), which distinguishes the type variable 'a from
the type variable ''a when compared with Tyvar.sameName.  Thus, we now
treat 'a, ''a, '''a, ... as distinct type variables (all but the first
having equality status).

Changed Tyvar.string, Tyvar.newString, and Tyvar.newNoname to
accomodate new representation.

SVN r7565