Use bit operations for real classification and low-level opeartions.
Use Real{32,64}.cast{to,from}Word primitives for implementation of
Real{32,64}.class and Real{32,64}.signBit.

Real{32,64}.{to,from}Decimal should respect sign of nan values.

Eliminates the need for C implementations of:
 IEEEReal.FloatClass.FP_{INFINITE,NAN,NORMAL,SUBNORMAL,ZERO}
 Real{32,64}.class (fpclassify)
 Real{32,64}.signBit (signbit)
 Real{32,64}.maxFinite
 Real{32,64}.minPos
 Real{32,64}.minNormalPos

There is a minor inconsistency in the Standard ML Basis Library
specification.  IEEEReal.toString states:
  "If the sign field is true, a #"~" is prepended."
This implies that a negative nan is converted to "~nan".
IEEEReal.toString also states:
  "The composition toString o REAL.toDecimal is equivalent to REAL.fmt StringCvt.EXACT."

However, REAL.{fmt,toString} states:
  "In all cases, ... NaN values are converted to the string "nan"."
This implies that a negative nan is converted to "nan".

This inconsistency is resolved in the implementation of the Basis
Library by breaking the equivalence of
"IEEEReal.toString o REAL.toDecimal" and "REAL.fmt StringCvt.EXACT"
for the case of a negative nan.  The former will return "~nan" (which
is rationalized by the explicit sign field of the
IEEEReal.decimal_approx type) and the latter will return "nan".

SVN r7573