Introduce an MLB annotation to control overload and flex record resolution scope.
The following MLB annotation is used to control the scope at which
overload constraints are resolved to default types (if not otherwise
resolved by type inference) and the scope at which unresolved flexible
record constraints are reported:

  resolveScope {strdec|dec|topdec|program}

The syntactic-class argument means to perform resolution checks at the
smallest enclosing syntactic form of the given class.  The default
behavior (corresponding to MLton's previous behavior) is to resolve at
the smallest enclosing strdec (which is equivalent to the largest
enclosing dec).  Other useful behaviors are to resolve at the smallest
enclosing topdec (which is equivalent to the largest enclosing strdec)
and at the smallest enclosing program (which corresponds to a single
.sml file and does not correspond to the whole .mlb program).

SVN r7618