Fixed bug in elaboration of strdescs in specs.
Thanks to Bernard Berthomieu for the bug report and example.

Fixed bug in elaboration that erroneously accepted the following:
    signature S = sig structure A : sig type t end
                            and B : sig type t end where type t = A.t end

The elaboration of a strdesc used a single Vector.foreach to elaborate
each sigexp and to bind each strid (to the result of the elaborated
sigexp); however, by binding each strid in the environment, the
earlier strid-s were in scope for later sigexp-s.

Changed the single Vector.foreach to a Vector.map to elaborate each
sigexp followed by a Vector.foreach to bind each strid (to the result
of the elaborated sigexp).

SVN r7619