Update .gitignore files.
Added leading '/' to most patterns in .gitignore files, in order to
limit the pattern to files in the same directory as the corresponding
.gitignore file.  Exceptions are '*.o' and '*.a' patterns in
'runtime/.gitignore' (in order to exclude object and archive files
anywhere in the 'runtime/' directory hierarchy) and '.cm/' patterns in
'lib/.gitignore', 'mlton/.gitignore', and 'util/cm2mlb/.gitignore' (in
order to exclude SML/NJ derived files).  Also added trailing '/' to
some patterns in .gitignore files, in order to limit the pattern to
directories.