Preliminary support for amd64 architecture with llvm codegen.
Array indexing on a 64-bit platform is done with 64-bit arithmetic,
which necessitated a change to the 'mkgep' function to take the size
of the offsets.

CPointer_toWord and CPointer_fromWord should bitcast with %uintptr_t
(alias of %Word32/i32 on 32-bit platforms and %Word64/i64 on 64-bit
platforms).

With these changes, all regressions except for exn-history{,3} pass
with the llvm codegen on amd64-darwin.