Perform cast after load in translation of Machine IL 'Cast' operand.
A 'Cast' operand will always be an r-value.  Hence, it is appropriate
to fetch the sub-operand as a value and cast the result.  In
particular, this ensures that no Machine IL registers are bitcast
(rather, the contents of the registers are bitcast), which in turn
allows the 'mem2reg' optimization of LLVM's 'opt' to completely
eliminate all 'alloca's that provide storage for Machine IL registers.