Translate IntInf consts to Big and Small representations.
Translate IntInf consts to Big and Small representations in conversion
from SSA to RSSA.  This is consistent with the treatment of other
IntInf operations in the conversion from SSA to RSSA: the
IntInf_toVector, IntInf_toWord, Word_toIntInf, and WordVector_toIntInf
primitives are translated to simple casts and other IntInf_*
primitives are translated to C calls.  Hence, after the conversion
from SSA to RSSA, IntInf is no longer treated as a primitive.