Update the WordXVector.layout function.
Update the WordXVector.layout function.  If the WordXVector's
elementSize is WordX.word8 and more than 90% of the characters satisfy
Char.isPrint, then display the WordXVector as an SML string constant
(with non-printable characters SML-escaped).  Otherwise, display the
WordXVector as an SML/NJ-style "#[0x0, 0xF]" vector literal.

The layout of large vector literals is not ideal, as the Layout
library (via the internal `sequence` function) with either place all
of the elemnts on the same line or all of the elements on separate
lines, whereas a "paragraph-fill" style would be preferable for large
vector literals.