Update model-elimination.sml benchmark program and iteration count.
Eliminate uses of `CommandLine.arguments ()` from model-elimination
benchmark program.  When executed by the benchmark program, the
model-elimination program is executed with no command-line arguments.
However, when executed by the update-counts.sh script, the
model-elimination program is executed with an interation count
command-line argument, which led to a Fail exception and an erroneous
iteration count.