Update Debian packaging (control, *.install)
First, usr/lib/mlton/opcodes was part of the discontinued bytecode
codegen. It doesn't exist anymore, so don't try to install it.

Second, the GNU architecture has changed from i486 to i586. I don't
know why the GNU architecture (which has changed twice, and may change
again to i686, for all I know) is used in all these paths instead of the
Debian architecture (which will presumably remain i386 until the end of
time). But I will assume there is a good reason, and so switch all
the paths to the new architecture.