Deterministically generate magic constant.
Each MLton generated executable embeds a magic constant that is meant
to protect against using MLton.World.load (or @MLton load-world
_world_) with a world saved by a different executable.

Previously, the magic constant was generated with MLton.Random.useed,
which uses /dev/urandom to get a random word, falling back to
String.hash of the input file name if reading from /dev/urandom fails
or if MLton.Random is provided by a stub.  Unfortunately, this means
that two builds of the same source file (by the same version of MLton)
will generate executables that differ in the value of the magic
constant; furthermore, this is the only difference in the two
executables (excluding the embedding of the output executable name in
the compile-time specified "@MLton" arguments; see 07afa54).

The magic constant is now generated by combining "String.hash
Version.banner" (which contains the version, build date, and build
node of the compiler) and "Random.word ()" (which is a deterministic
pseudo-random number generator).  Note that "Random.word" is used
throughout the compile to generate hash-values for identifiers and
other atomic constructs, so the state of the pseudo-random number
generator at the end of the compile is quite sensitive to the input
program.

Partly fixes MLton/mlton#50.