Add support for setting CM anchor bindings in cm2mlb tool
The cm2mlb "-map file" option allows setting a mapping from CM anchors
to MLB paths in order to generate MLB files with MLB paths that
correspond to CM anchors.

However, it is sometimes necessary to specify a CM anchor environment
in order for CM.Graph.graph to properly process a .cm file.  For
example:

devel.cm:
Library
  structure Lib
is
  sources.cm (bind:(anchor:util value:wip/util))

sources.cm:
Library
  structure Lib
is
  $util/sources.cm
  lib.sml

In order to process sources.cm with cm2mlb, a binding for $util is
necessary for CM.Graph.graph to process.  While a binding for $util is
provided by devel.cm, processing devel.cm with cm2mlb will simply
result in a MLB file with a reference to sources.mlb.

The cm2mlb "-bind file" option allows setting a binding from CM
anchors to CM paths.  Often, a "-bind file" should be accompanied by a
"-map file" in order to provide a CM anchor -> CM path environment for
CM.Graph.graph and to provide a CM anchor -> MLB path environment for
the resulting MLB file.