Always pretty print or-patterns with parens
Consider the following program:

  datatype t = T1 | T2 of int

  fun f1 t =
     case t of
        _ => true
      | T1 => false
      | T2 _ => false
      | T1 | T2 _ => false
      | (T1 | T2 _) | (T1 | T2 _) => false
      | x as (T1 | T2 _) | x as (T1 | T2 _) => false

  fun f2 l =
     case l of
        _ => true
      | [T1 | T2 _] => false

  fun f3 r =
     case r of
        _ => true
      | {t = T1 | T2 _} => false

The previous pretty printing of or-patterns would yield the following
warning messages:

  Warning: y.sml 5.4.
    Case has redundant rules.
      rules: T1 => false
             (T2 _) => false
             (T1 | T2 _) => false
             (T1 | T2 _ | T1 | T2 _) => false
             (x as T1 | T2 _ | x as T1 | T2 _) => false
      in: case t of _ => true | T1 => false  ...   T1 | T2 _) => false
  Warning: y.sml 14.4.
    Case has redundant rules.
      rules: [T1 | T2 _] => false
      in: case l of _ => true | [T1 | T2 _] => false
  Warning: y.sml 19.4.
    Case has redundant rules.
      rules: {t = T1 | T2 _} => false
      in: case r of _ => true | {t = T1 | T2 _} => false

Note that the fifth and sixth match rules of f1 are pretty printed
with insufficient parens.  With the current grammar/ast, it is not
trivial to detect the nesting of patterns, as a parenthesized pattern
is wrapped in a singleton Ast.Pat.FlatApp.

Now the pretty printing of or-patterns yields the following warning
messages:

  Warning: y.sml 5.4.
    Case has redundant rules.
      rules: T1 => false
             (T2 _) => false
             (T1 | T2 _) => false
             ((T1 | T2 _) | (T1 | T2 _)) => false
             (x as (T1 | T2 _) | x as (T1 | T2 _)) => false
      in: case t of _ => true | T1 => false  ...  T1 | T2 _)) => false
  Warning: y.sml 14.4.
    Case has redundant rules.
      rules: [(T1 | T2 _)] => false
      in: case l of _ => true | [(T1 | T2 _)] => false
  Warning: y.sml 19.4.
    Case has redundant rules.
      rules: {t = (T1 | T2 _)} => false
      in: case r of _ => true | {t = (T1 | T2 _)} => false

Note that the fifth and sixth match rules of f1 are pretty printed
with sufficient parens, but the second match rules of f2 and f3 are
printed with extraneous parens.

In general, it is desirable to never pretty print with insufficient
parens, even at the expense of occasionally pretty printing with
extraneous parens.